Move SynchronizerNodeService to apps-common
[ci]

Need to be able to reuse that in scan as well for the traffic endpoint.
