+++ apps/sv/src/main/scala/org/lfdecentralizedtrust/splice/sv/onboarding/SynchronizerNodeReconciler.scala
@@ -14,6 +14,7 @@ import org.lfdecentralizedtrust.splice.environment.{RetryFor, RetryProvider, Spl
 import org.lfdecentralizedtrust.splice.sv.onboarding.SynchronizerNodeReconciler.SynchronizerNodeState
 import org.lfdecentralizedtrust.splice.sv.store.SvDsoStore
 import org.lfdecentralizedtrust.splice.store.DsoRulesStore.DsoRulesWithSvNodeState
+import org.lfdecentralizedtrust.splice.sv.LocalSynchronizerNode
 import org.lfdecentralizedtrust.splice.sv.util.SvUtil
 import org.lfdecentralizedtrust.splice.sv.util.SvUtil.{LocalMediatorConfig, LocalSequencerConfig}
 import org.lfdecentralizedtrust.splice.util.PrettyInstances.*
@@ -23,7 +24,7 @@ import com.digitalasset.canton.topology.SynchronizerId
 import com.digitalasset.canton.tracing.TraceContext
 import com.digitalasset.canton.util.ShowUtil.*
 import org.lfdecentralizedtrust.splice.sv.config.SvScanConfig
-import org.lfdecentralizedtrust.splice.sv.SynchronizerNode.LocalSynchronizerNodes
+import org.lfdecentralizedtrust.splice.environment.SynchronizerNode.LocalSynchronizerNodes
 
 import scala.concurrent.{ExecutionContext, Future}
 import scala.jdk.CollectionConverters.MapHasAsScala
@@ -42,7 +43,7 @@ class SynchronizerNodeReconciler(
   private val dsoParty = dsoStore.key.dsoParty
 
   def reconcileSynchronizerNodeConfigIfRequired(
-      synchronizerNodes: Option[LocalSynchronizerNodes],
+      synchronizerNodes: Option[LocalSynchronizerNodes[LocalSynchronizerNode]],
       synchronizerId: SynchronizerId,
       state: SynchronizerNodeState,
       migrationId: Long,
